rework testing to move test to integration file
