+++ docs/reference/configuration/runtime-security/alerting.md
@@ -21,10 +21,6 @@ UDS Core ships Falco with [stable default rules](https://github.com/falcosecurit
 
 See more about default rules in the [Falco documentation](https://falco.org/docs/reference/rules/default-rules/).
 
-:::caution[Sandbox/Incubating Rules]
-UDS Core currently does not enable the Falco sandbox/incubating ruleset by default.
-:::
-
 #### Additional Rulesets
 
 :::note
@@ -52,6 +48,22 @@ This configuration:
 
 1. Enables the sandbox ruleset while excluding the "Terminal shell in container" rule
 2. Enables the incubating ruleset while excluding the "Read environment variable from /proc files" rule
+
+#### Finding Rule Names for `disabledRules`
+
+The rule names used in the `disabledRules` array should match the `rule` field from the Falco rules files. You can find these rule names in the following locations:
+
+1. **In the rule files shipped with UDS Core**:
+   - Sandbox rules: [`src/falco/chart/rules/sandbox-rules.yaml`](https://github.com/defenseunicorns/uds-core/blob/main/src/falco/chart/rules/sandbox-rules.yaml)
+   - Incubating rules: [`src/falco/chart/rules/incubating-rules.yaml`](https://github.com/defenseunicorns/uds-core/blob/main/src/falco/chart/rules/incubating-rules.yaml)
+
+   Look for entries that start with `- rule:` to find the rule names.
+
+2. **From Falco logs**:
+   When Falco detects an event, it logs the rule name in the output. You can find these logs by querying Loki with:
+   ```txt
+   {component="falco"} |~ "rule"
+   ```
 
 ### Querying Events with Loki
