+++ test/vitest/falco-integration.spec.ts
@@ -6,9 +6,9 @@
 import * as net from "net";
 import { afterAll, beforeAll, describe, expect, test } from "vitest";
 import { closeForward, getForward } from "./helpers/forward";
-import { execAndWait, withTempPod } from "./helpers/k8s";
-import { pollUntilSuccess } from "./helpers/polling";
+import { execAndWait, getAllLogsByLabelSelector, withTempPod } from "./helpers/k8s";
 import { queryLoki } from "./helpers/loki";
+import { pollUntilSuccess } from "./helpers/polling";
 
 describe("Falco Integration e2e Tests", () => {
   let lokiRead: { server: net.Server; url: string };
@@ -64,6 +64,52 @@ describe("Falco Integration e2e Tests", () => {
         expect(falcoEvent).toBeDefined();
         expect(falcoEvent!.status).toBe("success");
         expect(falcoEvent!.data.result.length).toBeGreaterThan(0);
+      },
+    );
+  }, 70000); // Set test timeout to 70 seconds
+
+  test("Falco detects 'Write below root' event and sends to Falco Sidekick", async () => {
+    // Generate a random string to identify this test run
+    const randomString = Math.random().toString(36).substring(2, 10);
+
+    // Use a temporary pod in kube-system to trigger the Falco event
+    await withTempPod(
+      {
+        name: `falco-write-test-${randomString}`,
+        namespace: "kube-system", // Use kube-system to get around zarf mutations
+        image: "alpine:latest",
+        command: ["sleep", "3600"],
+      },
+      async podName => {
+        // Try to write to a system directory (should trigger the rule)
+        await execAndWait(
+          "kube-system",
+          podName,
+          ["sh", "-c", `echo "test" >> /etc/passwd`],
+          "main",
+        );
+
+        // Poll for the Falco event in falcosidekick logs until success or timeout
+        const falcoSidekickEvent = await pollUntilSuccess(
+          async () => {
+            const falcoSidekickLogs = await getAllLogsByLabelSelector(
+              "falco",
+              "app.kubernetes.io/name=falcosidekick",
+            );
+            return falcoSidekickLogs.find(
+              log =>
+                log.includes('"rule":"Write below etc"') &&
+                log.includes("File below /etc opened for writing") &&
+                log.includes("file=/etc/passwd"),
+            );
+          },
+          result => result !== undefined,
+          "Falco event for write to /etc/passwd in falcosidekick logs",
+          60000, // 1 minute timeout
+          15000, // 15 seconds interval
+        );
+
+        expect(falcoSidekickEvent).toBeDefined();
       },
     );
   }, 70000); // Set test timeout to 70 seconds
